+++ src/com/intellij/idea/plugin/hybris/system/cockpitng/model/widgets/DynamicNode.java
@@ -22,10 +22,8 @@
 package com.intellij.idea.plugin.hybris.system.cockpitng.model.widgets;
 
 import com.intellij.idea.plugin.hybris.common.HybrisConstants;
-import com.intellij.util.xml.DomElement;
-import com.intellij.util.xml.GenericAttributeValue;
-import com.intellij.util.xml.Namespace;
-import com.intellij.util.xml.SubTagList;
+import com.intellij.idea.plugin.hybris.util.xml.SpringBeanReferenceConverter;
+import com.intellij.util.xml.*;
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -41,6 +39,7 @@ public interface DynamicNode extends DomElement, ExplorerNode {
      */
     @NotNull
     @com.intellij.util.xml.Attribute("populator-bean-id")
+    @Referencing(SpringBeanReferenceConverter.class)
     GenericAttributeValue<String> getPopulatorBeanId();